requirements.txt
apipkg==1.4
arrow==0.5.0
autopep8==1.2.4
bibtexparser==0.6.2
colorama==0.3.7
coverage==4.2
defusedxml==0.4.1
Django==1.10.1
django-autocomplete-light==3.1.8
django-mptt==0.8.6
django-polymorphic==1.0.1
djangorestframework==3.4.6
djangorestframework-gis==0.11
djangorestframework-recursive==0.1.1
execnet==1.4.1
future==0.14.3
GDAL==2.0.3
latexcodec==1.0.4
memoized-property==1.0.2
mendeley==0.3.2
numpy==1.12.0
oauthlib==0.7.2
pandas==0.19.2
pep8==1.7.0
psycopg2==2.6.2
py==1.4.31
pybib==2.2.1
pybtex==0.20.1
pytest==3.0.2
pytest-cache==1.0
pytest-cov==2.3.1
pytest-pep8==1.0.6
pytest-timeout==1.0.0
python-dateutil==2.5.3
python-termstyle==0.1.10
pytz==2016.10
PyYAML==3.12
requests==2.5.1
requests-oauthlib==0.4.2
six==1.10.0
suds-jurko==0.6
wstools==0.4.4
